+++ src/components/MainComponent.vue
@@ -1,9 +1,4 @@
 <script setup>
-import PlacesComponent from './PlacesComponent.vue';
-import MapComponent from './MapComponent.vue';
-import MapToggle from './MapToggle.vue';
-import PlacesModal from './PlacesModal.vue';
-import SearchBar from './SearchBar.vue';
 
 import { computed, ref, provide } from 'vue'
 import { ChevronDownIcon, EllipsisHorizontalIcon } from '@heroicons/vue/16/solid'
@@ -58,32 +53,19 @@ const transportationSwitch = computed(() => {
   return transportationOpen.value ? 'translate-x-[200%] transition-all' : 'translate-x-[300%]'
 })
 
-const isPlacesComponent = ref(true);
-
-const placesComponentCls = computed(() => {
-    return isPlacesComponent.value ? [""] : ["translate-y-full opacity-0"];
-});
-
 </script>
 
 <template>
 <div class="main-content">
-  <!-- 搜尋欄 -->
-  <div class="flex relative w-full item-center gap-2">
-    <SearchBar class="flex lg:w-[75%]  justify-end w-full  lg:ml-20"/>
-    <MapToggle 
-      class="hidden lg:absolute top-0 right-0  justify-start item-center mr-24 lg:flex"
-      v-model:isPlacesComponent="isPlacesComponent" />
-      <!-- 行程 toggle -->
+  
+  <!-- 行程 toggle -->
     <div 
     class="schedule  flex items-center  lg:absolute top-0 right-0 text-white w-[56px] hover:w-[86px] h-10 py-2 ps-3 pe-5 bg-yellow-500 rounded-s-full shadow-lg  hover:cursor-pointer" 
     @click="listToggle" 
     :class="scheduleToggleSwitch">
       <span class="inline-block w-6 h-6"><GlobeAsiaAustraliaIcon/></span>
       <p class="text-sm pl-0.5 font-medium hidden">行程</p>
-    </div>
   </div>
- 
 
   <div class="min-w-full h-screen flex flex-row-reverse bg-gray overflow-x-hidden">
     <div class="schedule-list w-full lg:w-96 h-screen bg-white shadow-xl transition-all" :class="listSwitch">
@@ -324,18 +306,6 @@ const placesComponentCls = computed(() => {
     <ScheduleDetail :class="detailSwitch"/>
     <TransportationWay :class="transportationSwitch"/> 
   </div>
-  <div>
-    <!-- 地圖/景點牆 -->
-    <div class="relative h-screen overflow-hidden">
-      <MapComponent />
-      <PlacesComponent class="absolute top-0 z-10 transition-all " :class="placesComponentCls" />
-      <PlacesModal class="hidden md:block"/>
-    </div>
-    <MapToggle 
-      class="absolute bottom-[1rem] left-[37%] z-20 justify-center item-center md:left-[44%] xl:hidden"
-      v-model:isPlacesComponent="isPlacesComponent" />
-  </div>
-  
 </div>
 </template>
